pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.5",
    "django-anymail>=13.0.1",
    "django-cors-headers>=4.7.0",
    "django-database-url>=1.0.3",
    "django-ratelimit>=4.1.0",
    "django-redis>=6.0.0",
    "django-restframework>=0.0.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.10",
    "gunicorn>=23.0.0",
    "psycopg-binary>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "redis>=6.4.0",
    "whitenoise>=6.9.0",
]
